Refactor 'Byte' class to common modules and update comments
