+++ task-framework/src/main/java/org/datatech/baikal/task/Framework.java
@@ -28,6 +28,7 @@ import org.apache.curator.utils.ZKPaths;
 import org.apache.zookeeper.CreateMode;
 import org.apache.zookeeper.KeeperException;
 import org.apache.zookeeper.data.Stat;
+import org.datatech.baikal.common.Bytes;
 import org.datatech.baikal.common.Configuration;
 import org.datatech.baikal.task.common.BaseTask;
 import org.datatech.baikal.task.common.DbType;
@@ -38,7 +39,6 @@ import org.datatech.baikal.task.dao.MonitorSchemaDao;
 import org.datatech.baikal.task.dao.MonitorTableDao;
 import org.datatech.baikal.task.dao.SourceDbDao;
 import org.datatech.baikal.task.processor.BaseTaskProcessor;
-import org.datatech.baikal.task.util.Bytes;
 import org.datatech.baikal.task.util.DtCmdUtil;
 import org.datatech.baikal.task.util.SystemContext;
 import org.datatech.baikal.task.util.ZkHandler;
