+++ task-framework/src/main/java/org/datatech/baikal/task/dao/SourceDbDao.java
@@ -17,9 +17,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.zookeeper.CreateMode;
+import org.datatech.baikal.common.Bytes;
 import org.datatech.baikal.task.common.DbType;
 import org.datatech.baikal.task.common.SourceDb;
-import org.datatech.baikal.task.util.Bytes;
 import org.datatech.baikal.task.util.ZkHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
