+++ task-framework/src/main/java/org/datatech/baikal/task/processor/SecondaryTaskProcessor.java
@@ -23,6 +23,7 @@ import org.apache.spark.launcher.SparkAppHandle;
 import org.apache.spark.launcher.SparkLauncher;
 import org.apache.zookeeper.CreateMode;
 import org.apache.zookeeper.data.Stat;
+import org.datatech.baikal.common.Bytes;
 import org.datatech.baikal.common.Configuration;
 import org.datatech.baikal.task.Config;
 import org.datatech.baikal.task.common.BaseTask;
@@ -34,7 +35,6 @@ import org.datatech.baikal.task.dao.EventLogDao;
 import org.datatech.baikal.task.dao.MetaConfigDao;
 import org.datatech.baikal.task.dao.SourceDbDao;
 import org.datatech.baikal.task.dao.SourceJdbcDao;
-import org.datatech.baikal.task.util.Bytes;
 import org.datatech.baikal.task.util.DtCmdUtil;
 import org.datatech.baikal.task.util.ZkHandler;
 import org.slf4j.Logger;
